Automatic versioning of JS and CSS, and improvement to Settings.
The AppController now computes MD5 sums for the JS and CSS files,
and add this to the query string, forces a cache drop when they change.
The MD5 sum is only computed once, and then stored in cache. This way,
the files are reloaded only when an updated is pushed through the CI
pipeline, and the cache is dropped.

The toolbar for CKEditor in the Settings page has been adjusted to
be more minimal.