Check the database user (not the Auth object) for admin privileges.
In UserController::login(), we checked the admin field of the user object given
by $this->Auth->user() to decide if the user should be sent to the admin page. This
is now empty when no unipi.ini configuration is present, and is only stored in the
database. Hence, we now use $this->user instead.