Add a whitelist for the filters to show in the badges.
The previous implementation of the filter badges elements would display all
query parameters in the page. Some of these may not be related to filters
in the page. Hence, a whitelist can now be optionally passed to filter
badges element to only show a subset of the query parameters, filtered
by their keys.